"Clone" the superagent for reuse
Allow the superagent to be cloned so it can be reused.
This allows you to set basic settings once (eg webserver startup), and then reuse the superagent.
Remove old "Timeout" method, and use the http.Client.Timeout instead.
"Safely" clone modifications the http.Client & Transport on a Clone if we can.